feat(ex00): date  and exhange_rate parsers implemented

+++ ex00/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: dcastor <dcastor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/08 09:21:53 by dcastor           #+#    #+#             */
-/*   Updated: 2025/09/08 16:41:34 by dcastor          ###   ########.fr       */
+/*   Updated: 2025/09/08 16:44:51 by dcastor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 class BitcoinExchange
 {
 private:
-	std::map<std::string, float> _data;
+	std::map<std::string, double> _data;
 
 	// Members functions
 	static std::string parseDate(const std::string &raw);
